image.php refactoring to support more media types
We send a max width and max height to image.php so that we don't end up
with previews which are too large and would slow down the experience.

image.php shows native SVGs and GIFs as is and generates a preview for
all the other supported media types.

Bonus: the link to the shared files work on OC7, but it's still a
security risk as it exposes all the files contained in the folder, not
just the images