Show pictures of all supported media types
Various preview providers can be turned on in the OC config and the
slideshow should display PNG previews of all files which can be
converted.
It should also add click handlers for all supported media types.

SVGs which are not converted are sent as "text/plain" as a security
measure and thus the browser can't show them.  The script now downloads
and base64 encodes them so that they can be used in img tags.